requirements.txt
# Core dependencies
streamlit==1.48.1
python-dotenv==1.1.1
requests==2.32.4

# Data processing
pandas==2.3.1
numpy==2.0.0

# AI/ML libraries
sentence-transformers==5.1.0
transformers>=4.55.2
torch==2.8.0
huggingface-hub==0.34.4

# Vector database
chromadb==1.0.17
faiss-cpu==1.12.0

# Workflow and LLM
langchain==0.3.27
langgraph==0.6.5
langchain-community==0.3.27
langchain-huggingface==0.3.1

# Utilities
pydantic==2.11.7
typing-extensions==4.14.1

# Logging
loguru>=0.7.3

# LLM
litellm==1.75.8
groq==0.31.0